Cosmetic changes

+++ auth/src/main/java/com/thex/chat/auth/dto/UserResponse.java
@@ -6,10 +6,10 @@ import com.thex.chat.auth.model.User;
 import java.util.Set;
 
 public record UserResponse(
-        Long id,
-        String username,
-        String email,
-        Set<Role> roles
+    Integer id,
+    String username,
+    String email,
+    Set<Role> roles
 ) {
     public static UserResponse from(User user) {
         return new UserResponse(user.getId(), user.getUsername(), user.getEmail(), user.getRoles());
